[AMap] Expose the consumer of AMapContext
