+++ src/AMap/Consumer.js
@@ -1,0 +1,3 @@
+import AMapContext from '../context/AMapContext';
+
+export default AMapContext.Consumer;
